These are libraries and files that will be used to make the driver for MySQL if need be.
I am unable at this moment to get Qt to send commands to the system.
The functionality is there and it should work. I think I am just
exhausted and might be running in circles. I think I might be able to
build the driver for us though. At this point, I am willing to write
in the data to a Sqlite Database for us to have something to show and
work with. The interface for using sqlite and mysqlite should be
pretty similar.

I feel like I am so close to gettting this to work.

To use the cmake and make functions you need to download cmake, GLib and
boost.